Separation of concerns: new decorators module abstraction
